Base Fragment
Base Fragment and Fragment Filter adapted to Base fragment
