+++ app/src/main/java/com/dmm/cheappcgames/ui/fragments/BaseFragment.kt
@@ -1,0 +1,39 @@
+package com.dmm.cheappcgames.ui.fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModel
+import androidx.viewbinding.ViewBinding
+import java.lang.IllegalArgumentException
+
+abstract class BaseFragment<VB: ViewBinding>(
+    private val bindingInflater: (inflater: LayoutInflater) -> VB
+) : Fragment() {
+
+    private lateinit var _binding: VB
+    protected val binding: VB get() = _binding
+
+    protected abstract fun onViewCreated()
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = bindingInflater.invoke(inflater)
+
+        if(_binding == null) { throw IllegalArgumentException("Binding null") }
+
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        onViewCreated()
+    }
+
+
+}
